refactor: customer dto 패키지 이동
